Remove Obsolete Types and Methods
